+++ jcli/src/jcli_lib/rest/v1/vote/mod.rs
@@ -1,5 +1,6 @@
 use crate::jcli_lib::rest::{Error, RestArgs};
 use crate::jcli_lib::utils::OutputFormat;
+use crate::utils::AccountId;
 use structopt::StructOpt;
 
 #[derive(StructOpt)]
@@ -11,14 +12,15 @@ pub struct AccountVotes {
     #[structopt(flatten)]
     output_format: OutputFormat,
 
-    /// Account address to filter votes
-    #[structopt(short, long)]
-    account: String,
+    /// Account id to filter votes.
+    /// An Account ID either in the form of an address of kind account, or an account public key.
+    #[structopt(short, long, parse(try_from_str = AccountId::try_from_str))]
+    account_id: AccountId,
 
     /// Id of the voteplan for which we want to list proposals
     /// the account voted for
     #[structopt(short, long)]
-    vote_plan_id: String,
+    vote_plan_id: Option<String>,
 }
 
 #[derive(StructOpt)]
@@ -38,19 +40,33 @@ impl Vote {
 
 impl AccountVotes {
     fn exec(self) -> Result<(), Error> {
-        let response = self
-            .args
-            .client()?
-            .get(&[
-                "v1",
-                "votes",
-                "plan",
-                &self.vote_plan_id,
-                "account-votes",
-                &self.account,
-            ])
-            .execute()?
-            .json()?;
+        let response = match self.vote_plan_id {
+            Some(vote_plan_id) => self
+                .args
+                .client()?
+                .get(&[
+                    "v1",
+                    "votes",
+                    "plan",
+                    &vote_plan_id,
+                    "account-votes",
+                    &self.account_id.to_url_arg(),
+                ])
+                .execute()?
+                .json()?,
+            None => self
+                .args
+                .client()?
+                .get(&[
+                    "v1",
+                    "votes",
+                    "plan",
+                    "account-votes",
+                    &self.account_id.to_url_arg(),
+                ])
+                .execute()?
+                .json()?,
+        };
         let formatted = self.output_format.format_json(response)?;
         println!("{}", formatted);
         Ok(())