+++ jormungandr-lib/src/interfaces/mod.rs
@@ -71,6 +71,6 @@ pub use self::transaction_witness::TransactionWitness;
 pub use self::utxo_info::{UTxOInfo, UTxOOutputInfo};
 pub use self::value::{Value, ValueDef};
 pub use self::vote::{
-    serde_base64_bytes, PrivateTallyState, Tally, TallyResult, VotePayload, VotePlan, VotePlanId,
-    VotePlanStatus, VotePrivacy, VoteProposalStatus,
+    serde_base64_bytes, AccountVotes, PrivateTallyState, Tally, TallyResult, VotePayload, VotePlan,
+    VotePlanId, VotePlanStatus, VotePrivacy, VoteProposalStatus,
 };