+++ jormungandr/src/rest/v1/handlers.rs
@@ -1,5 +1,5 @@
 use crate::rest::{v1::logic, ContextLock};
-use jormungandr_lib::interfaces::{Address, FragmentsBatch, VotePlanId};
+use jormungandr_lib::interfaces::{FragmentsBatch, VotePlanId};
 use warp::{reject::Reject, Rejection, Reply};
 
 impl Reject for logic::Error {}
@@ -40,13 +40,25 @@ pub async fn get_fragment_logs(context: ContextLock) -> Result<impl Reply, Rejec
         .map(|r| warp::reply::json(&r))
 }
 
-pub async fn get_account_votes(
+pub async fn get_account_votes_with_plan(
     vote_plan_id: VotePlanId,
-    account_id: Address,
+    account_id_hex: String,
     context: ContextLock,
 ) -> Result<impl Reply, Rejection> {
     let context = context.read().await;
-    logic::get_account_votes(&context, vote_plan_id, account_id)
+    logic::get_account_votes_with_plan(&context, vote_plan_id, account_id_hex)
+        .await
+        .map_err(warp::reject::custom)?
+        .ok_or_else(warp::reject::not_found)
+        .map(|r| warp::reply::json(&r))
+}
+
+pub async fn get_account_votes(
+    account_id_hex: String,
+    context: ContextLock,
+) -> Result<impl Reply, Rejection> {
+    let context = context.read().await;
+    logic::get_account_votes(&context, account_id_hex)
         .await
         .map_err(warp::reject::custom)?
         .ok_or_else(warp::reject::not_found)