+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - fix incorrect keys bech32 HRP by always using the ones provided by the library
+- update REST API: add new endpoint AccountVotes (/api/v1/votes/plan/account-votes/{account_id})
 
 ## Release 0.13.0